+++ src/plugins/monitors/monitors.c
@@ -215,7 +215,7 @@ monitor_set_foreground_color(Plugin *p, Monitor *m, const gchar *color)
 /******************************************************************************
  *                                 CPU monitor                                *
  ******************************************************************************/
-typedef unsigned long CPUTick;	   /* Value from /proc/stat                   */
+typedef unsigned long long CPUTick;/* Value from /proc/stat                   */
 typedef float CPUSample;	   /* Saved CPU utilization value as 0.0..1.0 */
 
 struct cpu_stat {
@@ -234,7 +234,7 @@ cpu_update(Monitor * c)
         FILE * stat = fopen("/proc/stat", "r");
         if (stat == NULL)
             return TRUE;
-        int fscanf_result = fscanf(stat, "cpu %lu %lu %lu %lu", 
+        int fscanf_result = fscanf(stat, "cpu %llu %llu %llu %llu",
                                     &cpu.u, &cpu.n, &cpu.s, &cpu.i);
         fclose(stat);
